Migrate money fields from Double to BigDecimal
Every currency-bearing field (Account.balance, Policy.coverageAmount/
premiumAmount, Payment.amount, Claim.amount, Advanced.loanAmount/
interestRate, Transaction.amount) moves from Double to BigDecimal with
@Column(precision = 19, scale = 2), across entities, DTOs, and the
deposit/withdraw arithmetic in AccountService. Double can't represent
decimals like 0.10 exactly, so repeated deposits/withdrawals could
silently drift by fractions of a cent — the classic float-for-money
bug. Beneficiary.percentage stays Double since it isn't currency.

Verified against the live local Postgres instance: Hibernate cleanly
altered all 8 affected columns from double precision to numeric(19,2)
against existing seeded data, and three successive $0.10 deposits
landed on exactly $100.80 with no drift (the same sequence under the
old Double implementation is the textbook case that goes wrong).

+++ src/main/java/com/eeluwole/finance_api/account/AccountController.java
@@ -7,6 +7,7 @@ import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -57,7 +58,7 @@ public class AccountController {
     @PatchMapping("/{id}/deposit")
     public ResponseEntity<AccountResponse> deposit(
             @PathVariable Long id,
-            @RequestParam Double amount,
+            @RequestParam BigDecimal amount,
             @AuthenticationPrincipal User currentUser) {
         return ResponseEntity.ok(accountService.deposit(id, amount, currentUser));
     }
@@ -66,7 +67,7 @@ public class AccountController {
     @PatchMapping("/{id}/withdraw")
     public ResponseEntity<AccountResponse> withdraw(
             @PathVariable Long id,
-            @RequestParam Double amount,
+            @RequestParam BigDecimal amount,
             @AuthenticationPrincipal User currentUser) {
         return ResponseEntity.ok(accountService.withdraw(id, amount, currentUser));
     }
