+++ src/main/java/com/eeluwole/finance_api/policy/dto/CreatePolicyRequest.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Data;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -20,11 +21,11 @@ public class CreatePolicyRequest {
 
     @NotNull
     @Positive
-    private Double coverageAmount;
+    private BigDecimal coverageAmount;
 
     @NotNull
     @Positive
-    private Double premiumAmount;
+    private BigDecimal premiumAmount;
 
     @NotNull
     private LocalDate startDate;
